+++ clang/test/AST/static-compound-literals-reeval.cpp
@@ -3,7 +3,6 @@
 struct RR { int r; };
 struct Z { int x; const RR* y; int z; };
 constinit Z z = { 10, (const RR[1]){__builtin_constant_p(z.x)}, z.y->r };
-// Check that we zero-initialize z.y->r.
+// Check that we zero-initialize z.y->r and that z.z sees the same value.
 // CHECK: @.compoundliteral = internal constant [1 x %struct.RR] zeroinitializer
-// FIXME: Despite of z.y->r being 0, we evaluate z.z to 1.
-// CHECK: global %struct.Z { i32 10, ptr @.compoundliteral, i32 1 }
+// CHECK: global %struct.Z { i32 10, ptr @.compoundliteral, i32 0 }